fix: properly depend on the stable test runner
