+++ packages/playwright-runner/package.json
@@ -1,6 +1,6 @@
 {
   "name": "playwright-runner",
-  "version": "0.3.2",
+  "version": "0.3.5",
   "license": "Apache-2.0",
   "author": {
     "name": "Microsoft Corporation"
@@ -12,7 +12,7 @@
   "description": "A test runner for running tests with Playwright.",
   "main": "out/index.js",
   "dependencies": {
-    "@playwright/test-runner": "0.3.2",
+    "@playwright/test-runner": "0.3.5",
     "playwright": "^1.3.0"
   }
 }
