+++ packages/test-runner/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@playwright/test-runner",
-  "version": "0.3.4-post",
+  "version": "0.3.5",
   "bin": {
     "test-runner": "./cli.js"
   },
@@ -33,7 +33,7 @@
     "text-diff": "^1.0.1"
   },
   "devDependencies": {
-    "@playwright/test-runner": "^0.2.11",
+    "@playwright/test-runner": "./stable-test-runner/node_modules/@playwright/test-runner",
     "@types/micromatch": "^4.0.1"
   },
   "license": "Apache-2.0",
