+++ stable-test-runner/package.json
@@ -1,0 +1,6 @@
+{
+  "private": true,
+  "dependencies": {
+    "@playwright/test-runner": "^0.3.4"
+  }
+}
